Split __index__ into modules, another code style pass
