+++ google_measurement_protocol/report.py
@@ -1,0 +1,31 @@
+import requests
+
+TRACKING_URI = 'https://ssl.google-analytics.com/collect'
+
+
+def report(
+        tracking_id, client_id, payloads, headers=None, **extra_info):
+    """Actually report measurements to Google Analytics."""
+    return [
+        _make_request(data, headers) for data in finalize_payloads(
+            tracking_id, client_id, requestable, **extra_info)]
+
+
+def _make_request(data, headers):
+    return requests.post(
+        TRACKING_URI, data=data, headers=extra_headers, timeout=5.0)
+
+
+def _finalize_payloads(tracking_id, client_id, payloads, **extra_info):
+    """Get final data for API requests for Google Analytics.
+
+    Updates payloads setting required non-specific values on data.
+    """
+    extra_payload = {
+        'v': '1', 'tid': tracking_id, 'cid': client_id, 'aip': '1'}
+
+    for payload in payloads:
+        final_payload = dict(payload)
+        final_payload.update(extra_payload)
+        final_payload.update(extra_info)
+        yield final_payload
